feat: Additional Search Bar Related Features
- The search page has a search input field. As the user types into the
search field, books that match the query are displayed on the page.
- Search results on the search page allow the user to select “currently
reading”, “want to read”, or “read” to place the book in a certain
shelf.
-When an item is categorized on the search page, and the user navigates
to the main page, it appears on that shelf in the main page.
-Books have the same state on both the search page and the main
application page: If a book is on a bookshelf, that is reflected in
both locations.